Refactored project structure
- Moved functionality to different packages
- Refactored components broken by move
