+++ src/main/java/DataTypes/CastlingRights.java
@@ -1,0 +1,8 @@
+package DataTypes;
+
+public enum CastlingRights {
+    QUEENSIDE,
+    KINGSIDE,
+    BOTH,
+    NONE
+}
