+++ src/main/java/DataTypes/Coordinate.java
@@ -1,0 +1,11 @@
+package DataTypes;
+
+public class Coordinate {
+    public Coordinate(int column, int row) {
+        this.column = (byte)column;
+        this.row = (byte)row;
+    }
+    public Coordinate() {}
+    public byte column;
+    public byte row;
+}
