+++ src/main/java/DataTypes/Move.java
@@ -1,22 +1,16 @@
-import java.util.*;
+package DataTypes;
 
-class Coordinate {
-    public Coordinate(int column, int row) {
-        this.column = (byte)column;
-        this.row = (byte)row;
-    }
-    public Coordinate() {}
-    public byte column;
-    public byte row;
-}
 
-class Move {
+public class Move {
+    public Coordinate origin;
+    public Coordinate destination;
+    public char promotion;
+
     public Move(Coordinate origin, Coordinate destination) {
         this.origin = origin;
         this.destination = destination;
         this.promotion = '-';
     }
-
     public Move(Coordinate origin, Coordinate destination, char promotion) {
         this.origin = origin;
         this.destination = destination;
@@ -30,21 +24,4 @@ class Move {
     }
 
     public Move() {}
-    public Coordinate origin;
-    public Coordinate destination;
-    char promotion;
-}
-
-class Moves extends LinkedList<Move> {}
-
-enum PlayerTurn {
-    WHITE,
-    BLACK,
-}
-
-enum CastlingRights {
-    QUEENSIDE,
-    KINGSIDE,
-    BOTH,
-    NONE
 }
