+++ src/main/java/DataTypes/Moves.java
@@ -1,0 +1,5 @@
+package DataTypes;
+
+import java.util.LinkedList;
+
+public class Moves extends LinkedList<Move> {}
