+++ src/main/java/DataTypes/PlayerTurn.java
@@ -1,0 +1,6 @@
+package DataTypes;
+
+public enum PlayerTurn {
+    WHITE,
+    BLACK,
+}
