+++ src/main/java/Main/Program.java
@@ -1,5 +1,6 @@
+package Main;
+
 import java.io.*;
-import java.util.*;
 import java.util.concurrent.*;
 
 public class Program {
