+++ src/test/java/UtilsTest.java
@@ -1,3 +1,4 @@
+import CommonTools.Utils;
 import org.junit.jupiter.api.*;
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
 public class UtilsTest {
